The api authentication are implemented

+++ BackEnd/app/Http/Controllers/API/AuthController.php
@@ -13,7 +13,7 @@ class AuthController extends Controller
     public function login(Request $request)
     {
         $credentials = Validator::make($request->all(), [
-            'username' => 'required|alpha',
+            'email' => 'required|email',
             'password' => 'required',
         ]);
 
@@ -24,7 +24,7 @@ class AuthController extends Controller
                 'errors' => $credentials->errors()->all(),
             ], 401);
         } else {
-            if (Auth::attempt(['name' => $request->username, 'password' => $request->password])) {
+            if (Auth::attempt(['email' => $request->email, 'password' => $request->password])) {
                 $user = $request->user();
                 return response()->json([
                     'status' => true,
@@ -58,7 +58,7 @@ class AuthController extends Controller
             ], 401);
         } else {
             $user = User::create([
-                'name' => $request->username,
+                'name' => $request->name,
                 'email' => $request->email,
                 'password' => bcrypt($request->password),
             ]);
@@ -76,14 +76,21 @@ class AuthController extends Controller
             }
         }
     }
-    public function logout(Request $request)
+    public function logout()
     {
         $user = Auth::user();
-        $user->tokens()->delete();
-        return response()->json([
-            'status' => true,
-            'message' => 'Logged Out Successful!',
-            'user' => $user,
-        ]);
+        if ($user) {
+            $user->tokens()->delete();
+            return response()->json([
+                'status' => true,
+                'message' => 'Logged Out Successful!',
+                'user' => $user,
+            ]);
+        } else {
+            return response()->json([
+                'status' => false,
+                'message' => 'user not found!',
+            ]);
+        }
     }
 }
